float -> int conversion

+++ engine/src/ceterrain.c
@@ -384,8 +384,8 @@ ce_scenenode* ce_terrain_find_scenenode(ce_terrain* terrain, float x, float z)
 	// FIXME: opengl hard-code
 	z = fabsf(z);
 
-	size_t sector_x = (size_t)x / (CE_MPRFILE_VERTEX_SIDE - 1);
-	size_t sector_z = (size_t)z / (CE_MPRFILE_VERTEX_SIDE - 1);
+	size_t sector_x = (int)x / (CE_MPRFILE_VERTEX_SIDE - 1);
+	size_t sector_z = (int)z / (CE_MPRFILE_VERTEX_SIDE - 1);
 
 	return terrain->scenenode->childs->items[sector_z *
 		terrain->mprfile->sector_x_count + sector_x];
